chore:댓글 화면 벗어나는 오류 해결
