+++ client/src/components/Community/Detail/Board.jsx
@@ -1,3 +1,5 @@
+// Board
+
 /* eslint-disable camelcase */
 import moment from 'moment';
 import React, { useEffect } from 'react';
