+++ client/src/components/Community/Detail/Comment.jsx
@@ -1,3 +1,6 @@
+// Comment
+// 댓글 삭제
+
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import Swal from 'sweetalert2';
@@ -97,7 +100,15 @@ const Comment = ({ comments }) => {
               <div className="ms-3">
                 {/* 댓글 작성자 정보 및 내용 */}
                 <div className="fw-bold">{comment.nickname || '사용자 없음'}</div>
-                <p style={{ whiteSpace: 'pre' }}>{comment.contents}</p>
+                <p
+                  style={{
+                    maxWidth: '500px',
+                    whiteSpace: 'pre-wrap',
+                    wordWrap: 'break-word',
+                    wordBreak: 'keep-all',
+                  }}>
+                  {comment.contents}
+                </p>
 
                 {/* 댓글 작성일 및 삭제 버튼 */}
                 <div className="mt-3">
