+++ client/src/components/Community/Detail/CommentInsertForm.jsx
@@ -1,3 +1,6 @@
+// CommentInsertFrom
+// 댓글 입력
+
 import React, { useState } from 'react';
 import { useRecoilCallback, useRecoilValue } from 'recoil';
 import { useParams, useNavigate } from 'react-router-dom';
