+++ client/src/components/Community/Detail/CommunityHeader.jsx
@@ -1,3 +1,5 @@
+// CommunityHeader
+
 import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
