+++ client/src/components/Community/Detail/CommunitySidebar.jsx
@@ -1,4 +1,5 @@
 // CommunitySidebar.jsx
+//  검색
 
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
